requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
duckduckgo_search==7.5.3
filelock==3.18.0
fsspec==2025.3.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
primp==0.14.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
six==1.17.0
smolagents==1.12.0
soupsieve==2.6
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
